requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.17
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djantic==0.3.4
dramatiq==1.17.1
gevent==24.11.1
greenlet==3.1.1
idna==3.10
pika==1.3.2
prometheus_client==0.21.1
psycopg2-binary==2.9.10
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
watchdog==6.0.0
watchdog_gevent==0.2.1
zope.event==5.0
zope.interface==7.2
# psycopg3 ?
PyYAML==6.0.2
pydantic==2.8.2
djantic==0.3.4
django-cors-headers==4.6.0

pytz==2025.1
python-dateutil==2.9.0.post0
